lcm/requirements.txt
# rest framework
Django==2.1.4
djangorestframework==3.9.4

# for access MySQL
PyMySQL==0.9.3

# redis cache
redis==2.10.5

# for access redis cache
django-redis-cache==0.13.1

# for call rest api
httplib2==0.12.3

# for unit test
coverage==4.2
mock==3.0.5
unittest_xml_reporting==1.12.0

# for auto-swagger
drf-yasg>=1.2.2
flex>=6.11.1
swagger-spec-validator>=2.1.0

# for onap logging
onappylog==1.0.9

# uwsgi for parallel processing
# uwsgi
